+++ ex01/srcs/Serializer.cpp
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   Serializer.cpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yabukirento <yabukirento@student.42.fr>    +#+  +:+       +#+        */
+/*   By: ryabuki <ryabuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/17 20:01:26 by yabukirento       #+#    #+#             */
-/*   Updated: 2025/05/17 20:10:03 by yabukirento      ###   ########.fr       */
+/*   Updated: 2025/06/11 13:54:28 by ryabuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/Serializer.hpp"
 
-uintptr_t Serializer::serialize(Data* ptr) {
-	return reinterpret_cast<uintptr_t>(ptr);
+unsigned long Serializer::serialize(Data* ptr) {
+	return reinterpret_cast<unsigned long>(ptr);
 }
 
-Data* Serializer::deserialize(uintptr_t raw) {
+Data* Serializer::deserialize(unsigned long raw) {
 	return reinterpret_cast<Data*>(raw);
 }
